requirements.txt
Flask==2.3.3
flask-jwt-extended==4.6.0
flask-limiter==3.5.1
cryptography==42.0.5
bcrypt==4.1.3
python-dotenv==1.0.1
